Safari-609.1.11.1/LayoutTests/workers/wasm-references-expected.txt

CONSOLE MESSAGE: line 1836: Finished test
PASS successfullyParsed is true

TEST COMPLETE
